# restaurants/specifications/swagger.yml
openapi: 3.0.1
info:
  title: Restaurants Service API
  description: Restaurants from GoOutSafe with Microservices
  version: "0.1"
servers:
- url: /

paths:
  /my_restaurant:
    post:
      summary: Get current logged owner's restaurant
      operationId: restaurants.resources.restaurants.my_restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
              required:
                - user_id
      responses:
        302: # TODO: change
          description: Redirects to /restaurants/details/<user_id>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        400:
          description: Owner id not linked to any restaurant
          content: {}
  
  /restaurants/{restaurant_id}:
    get:
      tags:
      - Restaurants
      summary: Get a single restaurant
      operationId: restaurants.resources.restaurants.restaurant_sheet
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Return restaurant
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - message
                  - restaurant_sheet
                properties:
                  resturant_sheet:
                    type: object
                    required:
                      - restaurant
                      - list_measures
                      - average_rate
                      - max_rate
                    properties:
                      restaurant:
                        $ref: '#/components/schemas/Restaurant'
                      list_measures:
                        description: String list of the security measures adopted by the restaurant
                        type: string
                      average_rate:
                        description: Average rate of the restaurant
                        type: number
                      max_rate:
                        description: Max rating of the restaurant
                        type: integer
                  status:
                    type: string
                  message:
                    type: string    
        400:
          description: Restaurant_id is not linked to any restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  
  /restaurants/like/{restaurant_id}:
    post:
      tags:
      - Restaurants
      summary: Toggle the like on a restaurant
      operationId: restaurants.resources.restaurants.like_toggle
      requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                required:
                  - user_id
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          type: integer
      responses:
        200:
          description: The like was correctly toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        400:
          description: User id is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        500:
          description: Error in toggling the like
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /restaurants/add/{id_op}:
    post:
      tags:
      - Restaurants
      summary: Creates a restaurant linked to the operator specified by id_op
      operationId: restaurants.resources.restaurants.post_add
      requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  address:          
                    type: string
                  city:
                    type: string
                  phone:          
                    type: string
                  menu_type:
                    type: string
                required:
                  - name
                  - address
                  - city
                  - phone
                  - menu_type
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Restaurant successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        400:
          description: Error in the data validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        500:
          description: Error in saving the restaurant in the db
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  
  /restaurants/details/{id_op}:
     get:
      tags:
        - Restaurants
      summary: Show the information needed to properly render the restaurant details page
      operationId: restaurants.resources.restaurants.details
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Returns all the restaurant information
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - message
                  - details
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
                    required:
                      - restaurant
                      - tables
                      - times
                      - avg_stay
                      - list_measures
                    properties:
                      restaurant:
                        $ref: '#/components/schemas/Restaurant'
                      tables:
                        description: List of the restaurant's tables
                        type: array
                        items:
                          $ref: '#/components/schemas/Table'
                      times:
                        description: The restaurant opening hours
                        type: array
                        items:
                          $ref: '#/components/schemas/RestaurantAvailability'
                      avg_stay:
                        description: The restaurant average stay time
                        type: string
                      list_measures:
                        description: String list of the security measures adopted by the restaurant
                        type: string
        400:
          description: The operator has no restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /restaurants/add_tables/{id_op}/{rest_id}:
    post:
      tags:
        - Restaurants
      summary: Adds tables to the restaurants 
      operationId: restaurants.resources.restaurants.add_tables
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      - name: rest_id
        in: path
        description: Restaurant's unique identifier
        required: true
        schema:
          type: integer
      requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  number:
                    type: integer
                  max_capacity:          
                    type: integer
                required:
                  - number
                  - max_capacity
      responses:
        200:
          description: Table successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        400:
          description: Value Error in the tables' capacities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  
  /restaurants/save_time/{id_op}/{rest_id}:
    post:
      tags:
        - Restaurants
      summary: Adds availabilties to the restaurant specified by the rest_id
      operationId: restaurants.resources.restaurants.add_time
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      - name: rest_id
        in: path
        description: Restaurant's unique identifier
        required: true
        schema:
          type: integer
      requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  day:
                    type: string
                  start_time:          
                    type: string
                  end_time:
                    type: string
                required:
                  - day
                  - start_time
                  - end_time
      responses:
        200:
          description: Availability successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        400:
          description: Can't add new availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  
  /restaurants/save_measure/{id_op}/{rest_id}:
    post:
      tags:
        - Restaurants
      summary: Adds security measures to the restaurant specified by the rest_id
      operationId: restaurants.resources.restaurants.add_measure
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      - name: rest_id
        in: path
        description: Restaurant's unique identifier
        required: true
        schema:
          type: integer
      requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  measure:
                    type: string
                required:
                  - measure
      responses:
        200:
          description: Measure successfully added
        400:
          description: Can't add new measure
          content: {}
  
  /restaurants/save_avg_time/{id_op}/{rest_id}:
    post:
      tags:
        - Restaurants
      summary: Adds average stay time to the restaurant specified by the rest_id
      operationId: restaurants.resources.restaurants.add_avg_stay
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      - name: rest_id
        in: path
        description: Restaurant's unique identifier
        required: true
        schema:
          type: integer
      requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  hours:
                    type: string
                  minutes:
                    type: string
                required:
                  - hours
                  - minutes
      responses:
        200:
          description: Average stay time successfully added
        400:
          description: Can't add new average stay time
          content: {}

    post:
      tags:
        - Restaurants
      summary: Edits the restaurant specified by the rest_id  
      operationId: restaurants.resources.restaurants.post_edit_restaurant
      parameters:
      - name: id_op
        in: path
        description: Operator's Unique identifier
        required: true
        schema:
          type: integer
      - name: rest_id
        in: path
        description: Restaurant's unique identifier
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:          
                  type: string
                city:
                  type: string
                phone:          
                  type: string
                menu_type:
                  type: string
              required:
                - name
                - address
                - city
                - phone
                - menu_type
      responses:
        200:
          description: Restaurant correctly modified
        400:
          description: Can't modify restaurant
          content: {}

components:
  schemas:
    Restaurant:
      description: The restaurant model
      required:
      - id
      - name
      - address
      - city
      - lat
      - lon
      - phone
      - owner_id
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          readOnly: true
          example: 123
        name:
          maxLength: 100
          minLength: 1
          type: string
          description: Restaurant's name
          example: "My Pizza Restaurant"
        address:
          maxLength: 100
          minLength: 1
          type: string
          description: Restaurant's address
          example: "Via dei Panini, 3"
        city:
          maxLength: 100
          minLength: 1
          type: string
          description: Restaurant's city
          example: "Pisa"
        lat:
          type: number
          description: Restaurant Latitude
          example: 42.42
        lon:
          type: number
          description: Restaurant Longitude
          example: 42.42
        phone:
          type: string
          minLength: 5
          maxLength: 50
          description: Restaurant Phone Number
          example: "050123456"
        owner_id:
          type: integer
          description: Unique owner identifier
          example: 123
    
    Table:
      description: The restaurant's table model
      required:
        - id
        - restaurant_id
        - capacity
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          readOnly: true
          example: 123
        restaurant_id:
          type: integer
          description: Restaurant's Unique Identifier
          example: 123
        capacity:
          type: integer
          description: The table max capacity
          minimum: 1
          maximum: 15
          example: 6
    
    RestaurantAvailability:
      description: The restaurant opening time model
      required:
        - id
        - start_time
        - end_time
        - day
        - restaurant_id
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          readOnly: true
          example: 123
        start_time:
          type: string
          description: The opening time start time (hour)
          example: "09:00" # TODO: check if this includes seconds
        end_time:
          type: string
          description: The opening time end time (hour)
          example: "12:00" # TODO: check if this includes seconds
        day:
          type: string
          description: The opening time week day
          pattern: '(Mon|Tues|Wednes|Thurs|Fri|Satur|Sun)day'
          example: "Tuesday"
        restaurant_id:
          type: integer
          description: Restaurant's Unique Identifier
          example: 123

    Response:
      description: Standard json response template
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: 'Success'
        message:
          type: string
          example: 'Restaurant successfully added'